Updated ninja with new fields
Ninjas will now realize they have been shot and will stay dead to work with the shoot() method in grid.
The earlier shoot() method would ruin their families and the ninjas wont be able to go home for the holidays, once they have been shot. The shoot() method would replace their place on the grid with an empty member. But that EmptyMember still had the ninja's location, so these assholes will still kill you even if it was an empty cell or the ninja's move() method would still "move" a ninja so its like you shot a ninja's double, but ended up looking retarded.

These changes to the Ninja class will make it so the ninja will return a boolean that they have been shot. The grid now possesses a method that checks if the ninja is dead, if they are dead, they cant move or kill you or respawn on the grid at all.

-added boolean isShot
-added boolean method totallyGotShot()
-added void method gotShot()