Update Grid.java
Commented out the Grid version of the printGrid() method in the code for moving that was creating duplicate prints of the Grid on runtime. I left the printGrid() method in there, but it is currently unused. Technically it shouldn't even be in there, but I'm leaving it just in case taking it out breaks something for now.