Added resetSeeAll()
This method will reset everything that is seen to false. This fix address Isaac's issue with the see() method leaving grid members as seen, even tho they are no longer in range. 

This method  will be called from GameEngine which in turn is called by UI before the grid is printed.